+++ src/test/java/dk/kb/storage/facade/DsStorageFacadeTest.java
@@ -6,18 +6,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import dk.kb.storage.facade.DsStorageFacade;
+
 import dk.kb.storage.model.v1.DsRecordDto;
+import dk.kb.storage.model.v1.MappingDto;
 import dk.kb.storage.model.v1.RecordTypeDto;
 import dk.kb.storage.storage.DsStorageUnitTestUtil;
 import dk.kb.util.webservice.exception.InternalServiceException;
-import dk.kb.util.webservice.stream.ExportWriter;
-import dk.kb.util.webservice.stream.ExportWriterFactory;
-
 
 public class DsStorageFacadeTest extends DsStorageUnitTestUtil{
 
@@ -138,16 +134,16 @@ public class DsStorageFacadeTest extends DsStorageUnitTestUtil{
      *   The 4 record origin defined in yaml used the config
      *
     - name: origin.strategy.none   
-      update_strategy: NONE
+      updateStrategy: NONE
 
     - name: origin.strategy.all    
-      update_strategy: ALL
+      updateStrategy: ALL
 
     - name: origin_strategy_child    
-      update_strategy: CHILD
+      updateStrategy: CHILD
 
     - name: oritin.strategy.parent
-      update_strategy: PARENT
+      updateStrategy: PARENT
      *  
      * The 4 unittest has minor differences in assertions about that is updated
      *   
@@ -539,8 +535,49 @@ public class DsStorageFacadeTest extends DsStorageUnitTestUtil{
         DsStorageFacade.createOrUpdateRecord(c1_2);
 
     }
-
+    @Test    
+    public void testCreateAndUpdateMapping() throws Exception {
     
-
+        String refId="referenceid_unittest_id123";
+        try {
+           MappingDto mapping = DsStorageFacade.getMapping("does_not_exist");
+           assertNull(mapping);          
+        }
+        catch (Exception e) {
+            fail();    
+        }
+     
+        //Create
+        try {
+            MappingDto mapping = new MappingDto();
+            mapping.setReferenceId(refId);
+            mapping.setKalturaId("kaltura_unittest_id123");
+            DsStorageFacade.createOrUpdateMapping(mapping);                      
+         }
+         catch (Exception e) {
+           fail();    
+         }
+        
+        //update
+        try {
+            MappingDto mapping = new MappingDto();
+            mapping.setReferenceId(refId);
+            mapping.setKalturaId("kaltura_unittest_id1234"); //4 added
+            DsStorageFacade.createOrUpdateMapping(mapping);                      
+         }
+         catch (Exception e) {
+           fail();    
+         }
+        
+        //read updated value
+        
+        try {        
+            MappingDto mapping = DsStorageFacade.getMapping(refId);
+            assertEquals("kaltura_unittest_id1234",mapping.getKalturaId());
+         }
+         catch (Exception e) {
+           fail();    
+         }        
+    }
 
 }